requirements.txt
python-crontab==2.3.8
GitPython==2.1.11
python-etcd==0.4.5
